backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
python-multipart==0.0.12
pymupdf==1.24.11
python-docx==1.1.2
openai==1.55.0
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
psycopg2-binary==2.9.9
supabase==2.10.0
